+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 ### Bug fixes:
   * Documentation links to elements declared in "auxiliary" IDL sources are now resolved correctly
     in the generated C++, Java, and Swift code.
+  * Internal implementation classes for Lambda interfaces in Java are now marked with `@exclude`
+    JavaDoc tag.
 ### Breaking changes:
   * Added validation against name clashes between property declarations in IDL files.